+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -3,10 +3,13 @@ import { bitKeepWallet } from './bitKeepWallet/bitKeepWallet';
 import { bitskiWallet } from './bitskiWallet/bitskiWallet';
 import { braveWallet } from './braveWallet/braveWallet';
 import { coinbaseWallet } from './coinbaseWallet/coinbaseWallet';
+import { coreWallet } from './coreWallet/coreWallet';
 import { dawnWallet } from './dawnWallet/dawnWallet';
+import { frontierWallet } from './frontierWallet/frontierWallet';
 import { imTokenWallet } from './imTokenWallet/imTokenWallet';
 import { injectedWallet } from './injectedWallet/injectedWallet';
 import { ledgerWallet } from './ledgerWallet/ledgerWallet';
+import { mMaskWallet } from './metaMaskWallet/MMaskWallet';
 import { metaMaskWallet } from './metaMaskWallet/metaMaskWallet';
 import { mewWallet } from './mewWallet/mewWallet';
 import { okxWallet } from './okxWallet/okxWallet';
@@ -15,7 +18,9 @@ import { phantomWallet } from './phantomWallet/phantomWallet';
 import { rabbyWallet } from './rabbyWallet/rabbyWallet';
 import { rainbowWallet } from './rainbowWallet/rainbowWallet';
 import { safeWallet } from './safeWallet/safeWallet';
+import { safeheronWallet } from './safeheronWallet/safeheronWallet';
 import { tahoWallet } from './tahoWallet/tahoWallet';
+import { talismanWallet } from './talismanWallet/talismanWallet';
 import { tokenPocketWallet } from './tokenPocketWallet/tokenPocketWallet';
 import { trustWallet } from './trustWallet/trustWallet';
 import { unipassWallet } from './unipassWallet/unipassWallet';
@@ -27,21 +32,28 @@ import { zerionWallet } from './zerionWallet/zerionWallet';
 
 export {
   argentWallet,
+  bitKeepWallet,
   bitskiWallet,
   braveWallet,
   coinbaseWallet,
+  coreWallet,
   dawnWallet,
+  frontierWallet,
   imTokenWallet,
   injectedWallet,
   ledgerWallet,
   metaMaskWallet,
+  mMaskWallet,
   mewWallet,
   omniWallet,
   phantomWallet,
   rabbyWallet,
   rainbowWallet,
   safeWallet,
+  safeheronWallet,
   tahoWallet,
+  talismanWallet,
+  tokenPocketWallet,
   trustWallet,
   okxWallet,
   walletConnectWallet,